+++ .storybook/preview.js
@@ -1,5 +1,6 @@
 import "../src/assets/css/styles.css";
 import "../src/assets/theme/styles.scss";
+import "react-calendar/dist/Calendar.css";
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },