+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.10.0](https://github.com/cmsgov/macpro-ux-lib/compare/v1.9.1...v1.10.0) (2022-09-07)
+
+
+### Features
+
+* **Datefield:** add Datefield component to allow users to work with a date input/date picker ([2fca61c](https://github.com/cmsgov/macpro-ux-lib/commit/2fca61cdd160c2efc739c5b74285192b48f609aa))
+
+## [1.9.1](https://github.com/cmsgov/macpro-ux-lib/compare/v1.9.0...v1.9.1) (2022-09-07)
+
+
+### Bug Fixes
+
+* **general:** fix broken image sources and some type errors ([a019609](https://github.com/cmsgov/macpro-ux-lib/commit/a01960979ee36b04259a470949b6b23e8d0acd22))
+
+# [1.9.0](https://github.com/cmsgov/macpro-ux-lib/compare/v1.8.0...v1.9.0) (2022-09-07)
+
+
+### Features
+
+* **Storybook:** allow main branch hosting on github pages ([2f9da25](https://github.com/cmsgov/macpro-ux-lib/commit/2f9da256b8e2ab64ca8a497ceb58e90ff992f2dc))
+
 # [1.8.0](https://github.com/cmsgov/macpro-ux-lib/compare/v1.7.1...v1.8.0) (2022-08-31)